+++ custom_components/myfox/entities/entities_camera.py
@@ -1,6 +1,6 @@
 import logging
 
-from . import MyFoxAbstractCameraEntity
+from entity import MyFoxAbstractCameraEntity
 from ..devices import BaseDevice
 from ..coordinator.myfox_coordinator import (MyFoxCoordinator)
